+++ src/flamegraph/merge.rs
@@ -108,6 +108,7 @@ fn flow<'a, LI, TI>(
 
 pub(super) fn frames<'a, I>(
     lines: I,
+    suppress_sort_check: bool,
 ) -> quick_xml::Result<(Vec<TimedFrame<'a>>, usize, usize, usize)>
 where
     I: IntoIterator<Item = &'a str>,
@@ -127,12 +128,14 @@ where
             continue;
         }
 
-        if let Some(prev_line) = prev_line {
-            if prev_line > line {
-                return Err(quick_xml::Error::Io(io::Error::new(
-                    io::ErrorKind::InvalidData,
-                    "unsorted input lines detected",
-                )));
+        if !suppress_sort_check {
+            if let Some(prev_line) = prev_line {
+                if prev_line > line {
+                    return Err(quick_xml::Error::Io(io::Error::new(
+                        io::ErrorKind::InvalidData,
+                        "unsorted input lines detected",
+                    )));
+                }
             }
         }